pyproject.toml
[build-system]
requires = ["poetry-core>=1.9.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "doxie"
version = "0.1.0"
description = "Doxie - Python project scaffold"
authors = ["Doxie Authors"]
license = "MIT"
package-mode = true
packages = [
  { include = "doxie", from = "src" }
]

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
fastmcp = "^2.11.3"
sqlalchemy = "^2.0.43"
pymupdf = "^1.26.3"
alembic = "^1.16.4"
python-docx = "^1.2.0"
markdown = "^3.8.2"
whoosh = "^2.7.4"
pydantic = "^2.11.7"
click = "^8.2.1"
python-dotenv = "^1.1.1"
PyYAML = "^6.0.2"
pydantic-settings = "^2.6.1"
psycopg = { version = "^3.2.1", extras = ["binary"] }
atlassian-python-api = "^4.0.4"
apscheduler = "^3.11.0"
beautifulsoup4 = "^4.13.4"


[tool.poetry.group.dev.dependencies]
pytest = "^8.4.1"
pytest-asyncio = "^1.1.0"
black = "^25.1.0"
flake8 = "^7.3.0"
mypy = "^1.17.1"

[tool.poetry.scripts]
doxie-mcp = "doxie.mcp.server:main"

[tool.black]
line-length = 100
target-version = ["py311", "py312"]

[tool.isort]
profile = "black"
